Full Screen Probs wont work for you

+++ Mars Raiders V2/Mars Raiders/Mars Raiders/Classes/Menu/Menu System.cs
@@ -32,7 +32,7 @@ namespace Mars_Raiders
         private void Collision()
         {
             Vector2 Temp;
-            Int16 X, Y;
+            Int32 X, Y;
             Temp = Functions.GetMouseCoords();
             X = Convert.ToInt16(Temp.X);
             Y = Convert.ToInt16(Temp.Y);
